TaskMasterAPI/requirements.txt
fastapi
sqlalchemy
uvicorn
python-jose
python-multipart
passlib[bcrypt]
